Tidying up - Refactored game files - Renamed store.js to globalState.js for clarity
