+++ src/components/globalState.js
@@ -1,6 +1,6 @@
 import { reactive } from "vue";
 
-export const store = reactive({
+export const globalState = reactive({
   playerPosition: { x: 736, y: 770 },
   setPlayerPosition(x, y) {
     this.playerPosition = { x: x, y: y + 10 };
